requirements.txt
Django
psycopg2
django-environ
djangorestframework
drf-yasg